+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 
 * update ng-alain 1.0.0-beta.4
 * update ng-zorro-antd 0.7.0-beta.2
+* ADD: minitable isEditCell
+* FIX: minitable editCell
 
 ## 0.2.5